Feat: 삭제 시 '삭제하시겠습니까?' 메시지 창 출력기능 추가

+++ src/main/webapp/WEB-INF/views/upform.jsp
@@ -6,6 +6,12 @@
 <link href="/css/upform.css" rel="stylesheet" type="text/css">
 <meta charset="UTF-8">
 <title>운세 수정하기</title>
+<script type="text/javascript">
+function del() {
+  if (confirm("정말 삭제하시겠습니까?"))
+    list_ok.submit();
+ }
+</script>
 </head>
 <body>
 	<img class="bana-pic" src="/resources/img/bana-img-logo-symbol2.png" alt="BanaPresso" onclick="location.href='/board/list'" />
@@ -23,7 +29,9 @@
 			</div>
 			<div class="btn-box">
 				<button class="btn-upform" type="submit">수정</button>
-				<button class="btn-delete" type="button" onclick="location.href='delete?id=${board.id}'">삭제</button>
+				<button class="btn-delete" type="button" onclick="if(!confirm('정말로 삭제하시겠습니까?')){return false;} else {return location.href='delete?id=${board.id}'}">삭제</button>
+				
+				<%-- onclick="location.href='delete?id=${board.id}'" --%>
 			</div>
 		</form>
 	</div>
